Make a generic filter hook
